cambios dto y relaciones

+++ src/bootcamps/bootcamps.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateBootcampDto } from './dto/create-bootcamp.dto';
 import { UpdateBootcampDto } from './dto/update-bootcamp.dto';
 import { Repository } from 'typeorm';
@@ -13,10 +13,10 @@ export class BootcampsService {
     private bootcampRepository: Repository<Bootcamp>){
 
   }
-  create(payload: any) {
+  create(body: CreateBootcampDto) {
     //1. crear una instancia de una entity bootcamp
     //2. grabar esa instancia y retornar 
-    const newBootcamp=this.bootcampRepository.create(payload);
+    const newBootcamp=this.bootcampRepository.create(body);
     return this.bootcampRepository.save(newBootcamp)
   }
 
@@ -25,19 +25,32 @@ export class BootcampsService {
   }
 
   findOne(id: number) {
-    return this.bootcampRepository.findOneBy({id});
+    const bootcamp=
+    this.bootcampRepository.findOneBy({id})
+    if(!bootcamp){
+      throw new NotFoundException(`No existe el bootcamp con id ${id}`)
+    }else{
+      return bootcamp
+    }
   }
 
-  async update(id: number, payload: any) {
+  async update(id: number, body: UpdateBootcampDto) {
     //Encontrar bootcampo por id
     //Hacer cambios por update: agregar cambios del payload a la entidad hallada en el punto 1
     const updBootcamp= await this.bootcampRepository.findOneBy({id});
-    this.bootcampRepository.merge(updBootcamp,payload)
+    if(!updBootcamp){
+      throw new NotFoundException(`No se puede actuliazar el bootcamp con id ${id}`)
+
+    }
+    this.bootcampRepository.merge(updBootcamp,body)
     return this.bootcampRepository.save(updBootcamp)
   }
 
   async remove(id: number) {
     const delBootcamp= await this.bootcampRepository.findOneBy({id});
+    if(!delBootcamp){
+      throw new NotFoundException(`No se puede eliminar el bootcamp con id ${id}`)
+    }
     this.bootcampRepository.delete(delBootcamp)
     return delBootcamp
   }
